Selecting and locating files in the media library.
