+++ AppleMusic/App/AppleMusicApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct AppleMusicApp: App {
     var body: some Scene {
         WindowGroup {
-            LibraryView()
+            AppleMusicTabView()
         }
     }
 }
