+++ AppleMusic/TabView/MediaLibrary/MediaEditing/Cells/Cells.swift
@@ -1,0 +1,22 @@
+//
+//  Cells.swift
+//  AppleMusic
+//
+//  Created by Ренат Сафин on 03.04.2022.
+//
+
+import SwiftUI
+
+var cells: [Cell] = [
+    Cell(name: "Плейлисты", nameImage: "music.note.list"),
+    Cell(name: "Артисты", nameImage: "music.mic"),
+    Cell(name: "Альбомы", nameImage: "square.stack"),
+    Cell(name: "Песни", nameImage: "music.note"),
+    Cell(name: "Телешоу и фильмы", nameImage: "tv"),
+    Cell(name: "Видеоклипы", nameImage: "music.note.tv"),
+    Cell(name: "Жанры", nameImage: "guitars"),
+    Cell(name: "Сборники", nameImage: "person.2.crop.square.stack"),
+    Cell(name: "Авторы", nameImage: "music.quarternote.3"),
+    Cell(name: "Загружено", nameImage: "arrow.down.circle"),
+    Cell(name: "Домашняя коллекция", nameImage: "music.note.house")]
+
