+++ AppleMusic/TabView/Radio/RadioView.swift
@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  RadioView.swift
 //  AppleMusic
 //
 //  Created by Ренат Сафин on 07.03.2022.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct ContentView: View {
+struct RadioView: View {
     var body: some View {
         Text("Hello, world!")
             .padding()
@@ -16,6 +16,6 @@ struct ContentView: View {
 
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
-        ContentView()
+        RadioView()
     }
 }
