+++ AppleMusic/TabView/Search/SearchView.swift
@@ -1,20 +1,20 @@
 //
-//  MiniPlayer.swift
+//  SearchView.swift
 //  AppleMusic
 //
-//  Created by Ренат Сафин on 07.03.2022.
+//  Created by Ренат Сафин on 03.04.2022.
 //
 
 import SwiftUI
 
-struct MiniPlayer: View {
+struct SearchView: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
-struct MiniPlayer_Previews: PreviewProvider {
+struct SearchView_Previews: PreviewProvider {
     static var previews: some View {
-        MiniPlayer()
+        SearchView()
     }
 }
